refactor: rename `konst/colours.rs` to `colour.rs`
Rename `lyra/src/core/konst/colours.rs` to `colour.rs`.
